Re-organize directory structure for data_utils.
