+++ fluid/DeepASR/data_utils/augmentor/trans_mean_variance_norm.py
